refactor(formatter/sort-imports): Split sort_imports modules (#15189)
- before: sort_imports.rs
- after: sort_imports/mod.rs + 3 files
